Reorder house polygons to anticlockwise
